src/embedding_models/embed.py
from langchain_huggingface import HuggingFaceEmbeddings
from dotenv import load_dotenv

load_dotenv()
embedding = HuggingFaceEmbeddings(model_name='sentence-transformers/all-MiniLM-L6-v2')

text = "This is Embedding Models"

vector = embedding.embed_query(text)
print(str(vector))