requirements.txt
transformers
langchain
huggingface_hub
langchain-huggingface
transformers
sentence-transformers
python-dotenv
huggingface_hub[hf_xet]



